Blinking Game Made Using SparkAr Studio
This Project is a mobile game made using SparkAR Studio. The concept of the game was to make a character move(in this case jump) just with the player doing an action(Blinking). The Game was made using simple backgrounds, a 3 character with 3 different poses, background music and sound effects for the movement and when the game ends. Simple coding was used to to adjust the timings for the movement of the character and collision of the character into obstacles.
- Anurag Jones Kukatla(19voa43004)